+++ src/main/java/sssj/io/VectorStream.java
@@ -3,6 +3,6 @@ package sssj.io;
 
 public interface VectorStream extends Iterable<Vector> {
 
-  public int numVectors();
+  public long numVectors();
 
 }
